# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Herald.lol Gaming Analytics API
  description: |
    # Herald.lol Gaming Analytics Platform API
    
    The Herald.lol API provides comprehensive access to League of Legends and Teamfight Tactics analytics, 
    player data, team management, and gaming insights. Built for developers, analysts, and gaming applications.
    
    ## Features
    
    - **Gaming Analytics**: KDA analysis, CS/min tracking, Vision Score, Damage Share, Gold Efficiency
    - **Riot API Integration**: Seamless access to Riot Games data with rate limit management
    - **Team Management**: Team creation, player management, performance tracking
    - **Real-time Data**: Live match analysis and gaming statistics
    - **Multi-region Support**: Global LoL regions (NA, EUW, KR, JP, etc.)
    
    ## Performance Targets
    
    - **Analysis Speed**: <5 seconds for gaming analytics
    - **UI Response**: <2 seconds for dashboard loading
    - **Uptime**: 99.9% platform availability
    - **Scalability**: Support for 1M+ concurrent users
    
    ## Authentication
    
    Herald.lol uses multiple authentication methods:
    
    - **Bearer Tokens**: JWT tokens for user authentication
    - **API Keys**: For server-to-server integration
    - **OAuth 2.0**: Social login integration
    - **MFA**: Multi-factor authentication for sensitive operations
    
    ## Rate Limiting
    
    Rate limits are enforced per subscription tier:
    
    - **Free**: 60 requests/minute
...
    
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for user authentication
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Gaming-API-Key
      description: API key for server-to-server authentication
    
    OAuth2:
      type: oauth2
      description: OAuth2 authentication with multiple providers
      flows:
        authorizationCode:
          authorizationUrl: https://api.herald.lol/auth/oauth/authorize
          tokenUrl: https://api.herald.lol/auth/oauth/token
          scopes:
            read: Read access to gaming data
            write: Write access to gaming data
            admin: Administrative access
            analytics: Access to advanced analytics

  # Common schemas
  schemas:
...
    Error:
      type: object
      required:
        - error
        - gaming_platform
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        code:
          type: string
          description: Error code
        gaming_platform:
          type: string
          enum: [herald-lol]
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier

    # Gaming user schema
...
    TeamStats:
      type: object
      properties:
        team_id:
          type: integer
          enum: [100, 200]
        win:
          type: boolean
        first_blood:
          type: boolean
        first_tower:
          type: boolean
        first_dragon:
          type: boolean
        first_baron:
          type: boolean
        tower_kills:
          type: integer
        dragon_kills:
          type: integer
        baron_kills:
          type: integer
        total_gold:
          type: integer
        total_kills:
          type: integer

    # Timeline event
...
    GamingTeam:
      type: object
      properties:
        id:
          type: string
          description: Unique team identifier
        name:
          type: string
          description: Team name
        display_name:
          type: string
        tag:
          type: string
          description: Team tag (3-5 characters)
        region:
          type: string
          enum: [NA, EUW, EUNE, KR, JP, BR, LAN, LAS, OCE, RU, TR]
        game_type:
          type: string
          enum: [LoL, TFT]
        tier:
          type: string
          enum: [amateur, semi_pro, professional]
        members:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
        stats:
          $ref: '#/components/schemas/TeamStats'
        achievements:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Team member schema
    TeamMember:
      type: object
      properties:
        user_id:
          type: string
        summoner_name:
          type: string
        role:
          type: string
          enum: [top, jungle, mid, adc, support, coach, analyst, manager]
        position:
          type: string
          enum: [captain, player, substitute, staff]
        main_champions:
          type: array
          items:
            type: string
        joined_at:
          type: string
          format: date-time
        games_played:
          type: integer
        wins:
          type: integer

    # Analytics response schema
...
  parameters:
...
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
      description: Team identifier
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip

  # Common responses
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid gaming parameters provided"
            details: "Summoner name must be between 3-16 characters"
            gaming_platform: "herald-lol"
...
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Insufficient gaming permissions"
            details: "Premium subscription required for advanced analytics"
            gaming_platform: "herald-lol"
    NotFound:
      description: Not found - resource doesn't exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Gaming data not found"
            details: "Summoner not found in specified region"
            gaming_platform: "herald-lol"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-Gaming-Rate-Limit:
          schema:
            type: integer
          description: Rate limit per minute
        X-Gaming-Rate-Remaining:
          schema:
            type: integer
          description: Remaining requests
        X-Gaming-Rate-Reset:
          schema:
            type: integer
          description: Reset time (Unix timestamp)
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Gaming API rate limit exceeded"
            limit: 60
            remaining: 0
            tier: "free"
            gaming_platform: "herald-lol"
...
paths:
...
  /gaming/teams:
    get:
      tags:
        - Team Management
      summary: List teams
      description: Get list of gaming teams with optional filtering
      operationId: getTeams
      parameters:
        - name: region
          in: query
          schema:
            type: string
            enum: [NA, EUW, EUNE, KR, JP, BR, LAN, LAS, OCE, RU, TR]
          description: Filter by region
        - name: tier
          in: query
          schema:
            type: string
            enum: [amateur, semi_pro, professional]
          description: Filter by competitive tier
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: '#/components/schemas/GamingTeam'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  gaming_platform:
                    type: string
                    enum: [herald-lol]
    
    post:
      tags:
        - Team Management
      summary: Create new team
      description: |
        Create a new gaming team. Requires team management permissions.
        
        **Rate Limits**: 5 team creations per day for Pro tier, 1 per day for Premium
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - tag
                - region
                - game_type
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 50
                tag:
                  type: string
                  minLength: 2
                  maxLength: 5
                region:
                  type: string
                  enum: [NA, EUW, EUNE, KR, JP, BR, LAN, LAS, OCE, RU, TR]
                game_type:
                  type: string
                  enum: [LoL, TFT]
                tier:
                  type: string
                  enum: [amateur, semi_pro, professional]
                  default: amateur
                description:
                  type: string
                  maxLength: 500
            example:
              name: "Herald Champions"
              tag: "HC"
              region: "NA"
              game_type: "LoL"
              tier: "amateur"
              description: "Competitive League of Legends team"
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/GamingTeam'
                  gaming_platform:
                    type: string
                    enum: [herald-lol]
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /gaming/teams/{teamId}:
    get:
      tags:
        - Team Management
      summary: Get team details
      description: Retrieve detailed information about a specific team
      operationId: getTeam
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/GamingTeam'
                  gaming_platform:
                    type: string
                    enum: [herald-lol]
        '404':
          $ref: '#/components/responses/NotFound'

  # Subscription and account endpoints  
...